+++ ext/duckdb/prepared_statement.c
@@ -72,7 +72,7 @@ VALUE rbduckdb_prepared_statement_new(duckdb_connection con, duckdb_extracted_st
     TypedData_Get_Struct(obj, rubyDuckDBPreparedStatement, &prepared_statement_data_type, ctx);
 
     if (duckdb_prepare_extracted_statement(con, extracted_statements, index, &(ctx->prepared_statement)) == DuckDBError) {
-        rb_raise(eDuckDBError, "Fail to get DuckDB::PreparedStatement object from ExtractedStatements object");
+        rb_raise(eDuckDBError, "Failed to create DuckDB::PreparedStatement object.");
     }
     return obj;
 }
@@ -114,7 +114,7 @@ static VALUE duckdb_prepared_statement_execute(VALUE self) {
         if (p == NULL) {
             p = duckdb_prepare_error(ctx->prepared_statement);
         }
-        rb_raise(eDuckDBError, "%s", p ? p : "Failed to execute prepared statement");
+        rb_raise(eDuckDBError, "%s", p ? p : "Failed to execute prepared statement.");
     }
     return result;
 }